+++ docs/RELEASE-NOTES.rst
@@ -26,6 +26,8 @@ Bug Fixes
 * `Issue 3432 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3432>`_: Show meaningful logs for exceptions occured from controller agent
 * `Issue 3396 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3396>`_: Fix adding pool members from external clusters in nodeportLocal mc mode
 * `Issue 3351 <https://github.com/F5Networks/k8s-bigip-ctlr/issues/3351>`_: improve message handling when getting HTTP/401 from AS3
+* Fix pool members not getting updated for VS/TS on re-deployment of application with different servicePort and targetPort.
+
 
 Upgrade notes
 ``````````````